split up init.el and byte-compile configs
It had reached 2.5kloc. Configurations are now more or less split up in
some logical fashion. A lot of fixes were made along the way, including
fixes that were causing warnings. Some tweaks had to be made for the
files to byte-compile properly. May need to revisit this later, but
everything is working fine for now.