auto-fill only comments in prog-mode derived modes
Also, note that when comment-auto-fill-only-comments is set to non-nil,
other modes such as the one used in git commit buffers as well as
markdown-mode stop auto-filling, presumably because the regular text in
those modes are not comments and so nothing is auto-filled.

This fixes that by creating a function that locally sets it to true and
enables auto-fill mode, and this function is hooked onto the
prog-mode-hook.