f9 toggles scroll-bars
The scroll-bars can be useful for determining how large a buffer is and
how far into the buffer we are. I feel like this is better than
something cluttering up the mode-line, and I get a quicker and better
sense of location.

I configured the scroll-bars to appear on the left, though I may change
that later.