use yasnippet!
I avoided it at first but it seems like a very useful thing such as for
jsdoc insertion [0].

[0]: https://github.com/mooz/js-doc/pull/16

I don't want it to hijack my TAB key so I unbind that and bind yasnippet
expansion to M-n, and M-N triggers company-yasnippet.

I make sure that evil-insert state is set before expanding a snippet.