more robust comment-indent-new-line
It now works in Go and in text-mode derived modes such as the git commit message
buffers.

This foregoes the syntax table check to see if we were in a comment, since the
decision ultimately comes down to whether there's a comment-start right before
the point. This makes it more flexible for use in text-modes.

It would be nice if on cancellation, i.e. pressing ESC, the padded space were
removed. I believe this is done in Emacs itself but I can't remember where or
how.