change Xresources strategy
GDM runs xrdb with the -nocpp option which prevents the use of the
preprocessor for both defining colors as well as including other files.

In order to work around this, change the .Xresources strategy from one
of simply symlinking into one where the `set-theme` command is used to
concatenate it with the .Xresources.light or .Xresources.dark variant
and output the concatenation result into the ~/.Xresources location.

The caveat with this approach is that `set-theme` will need to be set
whenever changes are made to .Xresources, but the only other alternative
I see is modifying /etc/gdm/Xsession to remove the -nocpp option, but
the file is owned by gdm and is therefore a brittle solution.