don't use font-awesome on Windows
I finally tracked down the severe performance issues with Emacs on
Windows. The culprit was Windows' rendering of font-awesome. If a
font-awesome icon is inserted into a buffer, everything slows down
considerably. As soon as the icon is deleted, performance returns to
normal.

Instead use simple Unicode icons on Windows.