emacs: refactor themes
Previously we were loading the given solarized variant and then
overriding face definitions of the theme itself. However, this is
unnecessary because Emacs supports loading multiple themes and
"layering" them based on precedence, similar to CSS.

So now I created a new theme called solarized-ext which houses the
changes and additions to the solarized theme. Then when loading themes I
load solarized first, then solarized-ext, so that the latter takes
precedence.

This compartmentalizes changes to our own theme.

Now making changes to solarized-ext simply entails evaluating the
theme's buffer.

One thing to note is that because theme definitions are layered,
anything we don't want to inherit from solarized should be explicitly
disabled.

For example when I first made this change I was noticing that the
mode-line's padding section between center and right was not styling
correctly. This ended up being because solarized's mode-line definition
had a :box property which was applying to the padding section. To
resolve this I had to explicitly set :box unspecified on the face in
question. This also applied to header-line the face. More of these kinds
of issues may come to light over time.