emacs: add my-window-resizer
Bound to PRE w r.

Finally a very fast and, most importantly, intuitive, way to resize
windows. This command enables a transient-map which binds 'hjkl' to
resize the window. Window resizing is defined by moving the right or
bottom edge of the window. This only makes sense if there is another
window beyond the given edge. For example, resizing the right edge
doesn't make sense if there is no window to the right. I could have made
it so resizing in that case didn't do anything, but I found it more
useful to make it so it applies to the left edge instead.

If the universal argument is present, then resizing applies to the
opposing edge if there is a window beyond it. For example, instead of
'l' moving the right edge to the right, it moves the left edge to the
right provided that there is a window to the left.

Pressing C-u (universal argument) within this transient-map works as a
toggle, so one can press C-u so that all further resize operations apply
to the left or top edges.

Further still, it's possible to press 'w' to invoke ace-window to
seamlessly switch to another window and _continue_ to proceed to resize
that new window.

As with all transient maps, if a key is pressed that isn't covered by
it, the transient map is removed and the key passes-through and is
processed as usual. The 'q' or ',' keys can be used to explicitly exit
the transient map.

Transient-map Keys:

* j   - move the bottom edge down if possible, else the top edge down
* k   - move the bottom edge up if possible, else the top edge up
* h   - move the right edge left if possible, else the left edge left
* l   - move the right edge right if possible, else the left edge right
* C-u - make each operation apply to the top or left edge, if possible
* w   - ace-window to switch windows
* c   - close the window
* q   - explicitly quit the transient map
* ,   - explicitly quit the transient map
* J   - switch to the window below
* K   - switch to the window above
* H   - switch to the window to the left
* L   - switch to the window to the right

This also enables which-key's support for transient-maps in order to
show the available options while the transient-map is active.

This also unbinds frame-cmd's global C-M-S-{hjkl} binds and repurposes them to
the above commands:

* C-M-S-h - my-move-splitter-left
* C-M-S-l - my-move-splitter-right
* C-M-S-k - my-move-splitter-down
* C-M-S-j - my-move-splitter-up

Since PRE w r is now taken, winner-redo was bound to C-S-l and
winner-undo was also bound to C-S-h.