emacs: use Hydra
I previously didn't have a use for Hydra, but now that I learned about
transient maps I figure I'll reuse the infrastructure provided by Hydra
instead of doing things manually.

One thing I noticed was that when hydra-is-helpful is non-nil, which is
the default, repeating commands is slower and actually somehow breaks at
the extents. This means we can't show the hint popup, but we don't need
it anyway. Worst case we can probably turn on which-key's transient map
again.