emacs: add my-open-bookmark
Like my-open-{file,buffer}, when in a projectile project it'll only show
bookmarks on files that are part of the project. Otherwise, or if there
is a prefix argument, it'll show them all via helm-filtered-bookmarks.

Move the binding to "o m" as in "open mark" since the "b" prefix doesn't
make much sense.