emacs: refactor mode-line.el
Got rid of the mode-line-center since there was no point to it anyways,
nothing was being centered in the mode-line, it just acted as an
extension of mode-line-left.

Mode-line rendering function now takes an optional parameter specifying
the maximum width of an extra component that will be placed in between
left and right. This is used for the mode-line rendering (and not for
title-format or header-line) with the buffer-identification component.
The buffer-identification component is provided with the amount of space
it has to render itself, and if it wouldn't fit it chooses to only
display the filename as opposed to the full path.

The projectile project component was extracted from the
buffer-identification component.