emacs: global binds to scroll last-used window
I often find myself wanting this, and I already added it to
git-commit-mode and it was very useful. Instead of using
scroll-other-window, which determines the "other window" based on the
layout of the windows and so may not be what I expect, I wrote functions
to scroll the MRU window.