emacs: add my-helm-faces-at-point
This helm command lists all of the faces that apply to the character at
the point. It also shows right-aligned example text propertized with
that face.

The default action describes that face via describe-face. Other actions
can insert the name into the buffer or copy the name.

The key `C-c C-l` is bound to open a buffer with the output of
list-faces-display that is limited to only the faces that apply to the
point.