zsh: introduce hash_override
It can be used to override the paths to certain binaries, independent of
PATH resolution.

Note that it should be re-run after any changes to PATH, since changes
to PATH will invalidate and recompute the command hash table.

This is primarily useful on macOS after installing brew's GNU coreutils,
instead of outright replacing all of macOS's coreutils with brews, which
I'm unsure if I want to do yet.

The reason I'm uncertain is that I worry that certain programs might
detect macOS and specially-handle macOS's coreutils distribution, which
would then break if the GNU's coreutils are overriding macOS'.

Maybe this is an unnecessary concern and I should instead override with
GNU's coreutils and fix any issues that come up after doing that. I may
yet try that.